+++ asteca/cluster.py
@@ -12,7 +12,8 @@ class cluster:
     Parameters
     ----------
     cluster_df : pd.DataFrame
-        pandas DataFrame with the cluster's loaded data
+        `pandas DataFrame <https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.html>`_
+        with the cluster's loaded data
     magnitude : str
         Name of the DataFrame column that contains the magnitude
     e_mag : str
@@ -37,6 +38,7 @@ class cluster:
         Name of the DataFrame column that contains the second color's uncertainty
 
     """
+
     cluster_df: pd.DataFrame
     magnitude: str
     e_mag: str
@@ -59,7 +61,7 @@ class cluster:
         The photometry is store with a '_p' to differentiate from the self.magnitude,
         self.color, etc that are defined with the class is called.
         """
-        print("Reading and processing cluster data")
+        print("Instantiating cluster...")
 
         self.mag_p = np.array(self.cluster_df[self.magnitude])
         self.e_mag_p = np.array(self.cluster_df[self.e_mag])
@@ -75,6 +77,11 @@ class cluster:
             self.ra_v = self.cluster_df[self.ra]
             self.dec_v = self.cluster_df[self.dec]
 
+        print(f"N_stars        : {len(self.mag_p)}")
+        print(f"Magnitude      : {self.magnitude}")
+        print(f"Color          : {self.color}")
+        if self.color2 is not None:
+            print(f"Color2         : {self.color2}")
         print("Cluster object generated\n")
 
     def radecplot(self):
@@ -110,14 +117,17 @@ class cluster:
 
         return ax
 
-    def clustplot(self, ax=None, binar_prob=None):
+    def clustplot(self, ax, color_idx=0, binar_probs=None):
         r"""Generate a color-magnitude plot.
 
         Parameters
         ----------
         ax : matplotlib.axis, optional, default=None
             Matplotlib axis where to draw the plot
-        binar_prob : numpy.array, optional, default=None
+        color_idx : int, default=0
+            Index of the color to plot. If ``0`` (default), plot the first color. If
+            ``1`` plot the second color.
+        binar_probs : numpy.array, optional, default=None
             Array with probabilities of being a binary system for each observed star
 
         Returns
@@ -125,20 +135,23 @@ class cluster:
         matplotlib.axis
             Matplotlib axis object
 
-
         """
-        if ax is None:
-            f, ax = plt.subplots()
+        if color_idx > 1:
+            raise ValueError(
+                f"Wrong 'color_idx' value ({color_idx}), should be one of: [0, 1]"
+            )
 
-        if binar_prob is not None:
-            msk_binar = binar_prob > 0.5
+        if binar_probs is not None:
+            msk_binar = binar_probs > 0.5
 
         mag_col = self.magnitude
         col_col = self.color
+        if color_idx == 1:
+            col_col = self.color2
 
-        if binar_prob is None:
+        if binar_probs is None:
             ax.scatter(
-                self.colors_p[0],
+                self.colors_p[color_idx],
                 self.mag_p,
                 c="green",
                 alpha=0.5,
@@ -146,21 +159,25 @@ class cluster:
             )
         else:
             ax.scatter(
-                self.colors_p[0][~msk_binar],
+                self.colors_p[color_idx][~msk_binar],
                 self.mag_p[~msk_binar],
-                c="green",
+                # c="green",
+                c=binar_probs[~msk_binar],
+                marker="o",
                 alpha=0.5,
                 label=f"Observed (single), N={len(self.mag_p[~msk_binar])}",
             )
             ax.scatter(
-                self.colors_p[0][msk_binar],
+                self.colors_p[color_idx][msk_binar],
                 self.mag_p[msk_binar],
-                c="red",
+                # c="red",
+                c=binar_probs[msk_binar],
+                marker="s",
                 alpha=0.5,
                 label=f"Observed (binary), N={len(self.mag_p[msk_binar])}",
             )
 
-        ax.set_ylim(max(self.mag_p) + .5, min(self.mag_p) - .5)
+        ax.set_ylim(max(self.mag_p) + 0.5, min(self.mag_p) - 0.5)
         ax.set_xlabel(col_col)
         ax.set_ylabel(mag_col)
         ax.legend()