+++ asteca/modules/synth_cluster_priv.py
@@ -145,10 +145,12 @@ def add_binarity(self) -> np.ndarray:
         N_cols: magnitude + colors + initial mass
 
     """
-    print("Generating binary data")
+    all_colors = [self.isochs.color]
+    if self.isochs.color2 is not None:
+        all_colors.append(self.isochs.color2)
+    N_colors = len(all_colors)
 
     mag_idx = 0
-    N_colors = len(self.isochs.color_filter_name)
     m_ini_idx = mag_idx + N_colors + 1
 
     # Extend to accommodate binary data
@@ -175,7 +177,7 @@ def add_binarity(self) -> np.ndarray:
             theor_tracks[mx][ax][m_ini_idx + 1] = mag_combine(mag, mag_m2)
 
             # Calculate unresolved color for each color defined.
-            for ic, color in enumerate(self.isochs.color_filter_name):
+            for ic, color in enumerate(all_colors):
                 mc1 = self.isochs.color_filters[mx][ax][color[0]]
                 mc2 = self.isochs.color_filters[mx][ax][color[1]]
                 f_m1 = np.interp(m2, mass_ini, mc1)
@@ -190,32 +192,92 @@ def add_binarity(self) -> np.ndarray:
     return theor_tracks
 
 
-def extinction_coeffs(self) -> list:
+def ccmo_ext_coeffs(self) -> list:
     """Obtain extinction coefficients for all the observed filters and colors,
     in the order in which they are stored in theor_tracks.
 
     ext_coefs = [ec_mag, ec_col1, ...]
 
     """
-    print("Obtaining extinction coefficients")
-
-    # For the magnitude.
-    # Effective wavelength in Armstrong.
-    eff_wave = self.isochs.mag_color_lambdas[self.isochs.mag_filter_name]
-    # CCM coefficient. Effective wavelength in inverse microns.
-    ext_coefs = [ccm_model(10000.0 / eff_wave)]
-
-    # For colors.
-    for color in self.isochs.color_filter_name:
-        c_filt1, c_filt2 = color
+    ext_coefs = [[], []]
+    if self.ext_law != "GAIADR3":
+        # For the magnitude.
         # Effective wavelength in Armstrong.
-        eff_wave1 = self.isochs.mag_color_lambdas[c_filt1]
-        eff_wave2 = self.isochs.mag_color_lambdas[c_filt2]
-        ext_coefs.append(
-            [ccm_model(10000.0 / eff_wave1), ccm_model(10000.0 / eff_wave2)]
-        )
+        eff_wave = self.isochs.magnitude_effl
+        # Effective wavelength in inverse microns.
+        ext_coefs[0] = ccmo_model(10000.0 / eff_wave)
+
+        # For colors.
+        eff_wave1, eff_wave2 = self.isochs.color_effl
+        ext_coefs[1] = [ccmo_model(10000.0 / eff_wave1), ccmo_model(10000.0 / eff_wave2)]
+
+    if self.isochs.color2_effl is not None:
+        eff_wave1, eff_wave2 = self.isochs.color2_effl
+        ext_coefs += [[ccmo_model(10000.0 / eff_wave1), ccmo_model(10000.0 / eff_wave2)]]
 
     return ext_coefs
+
+
+def ccmo_model(mw: float) -> list:
+    """Cardelli, Clayton, and Mathis (1989 ApJ. 345, 245) model for extinction
+    coefficients with updated coefficients for near-UV from O'Donnell
+    (1994, ApJ, 422, 158).
+
+    ccm_coef = a + b / Rv
+
+    Implementation taken from:
+
+    http://idlastro.gsfc.nasa.gov/ftp/pro/astro/ccm_unred.pro
+
+    There appears to be an error in the Far-UV range in the original IDL
+    routine where the maximum inverse wavelength is 11 and it should be 10
+    according to Cardelli et al. 1989 (pag 251, Eq (5,a,b)).
+
+    """
+
+    if 0.3 <= mw < 1.1:
+        # Infrared.
+        a, b = 0.574 * (mw**1.61), -0.527 * (mw**1.61)
+
+    elif 1.1 <= mw < 3.3:
+        # Optical/NIR.
+        # Original coefficients from CCM89
+        # c1 = [1., 0.17699, -0.50447, -0.02427, 0.72085, 0.01979, -0.77530,
+        #       0.32999]
+        # c2 = [0., 1.41338, 2.28305, 1.07233, -5.38434, -0.62251, 5.30260,
+        #       -2.09002]
+        # New coefficients from O'Donnell (1994)
+        c1 = [1.0, 0.104, -0.609, 0.701, 1.137, -1.718, -0.827, 1.647, -0.505]
+        c2 = [0.0, 1.952, 2.908, -3.989, -7.985, 11.102, 5.491, -10.805, 3.347]
+        y = mw - 1.82
+        # Reverse because polyval starts from the highest degree.
+        c1.reverse(), c2.reverse()
+        a, b = np.polyval(c1, y), np.polyval(c2, y)
+
+    elif 3.3 <= mw < 8.0:
+        # Mid-UV
+        F_a, F_b = 0.0, 0.0
+        if mw >= 5.9:
+            y = mw - 5.9
+            F_a = -0.04473 * y**2 - 0.009779 * y**3
+            F_b = 0.2130 * y**2 + 0.1207 * y**3
+        a = 1.752 - 0.316 * mw - (0.104 / ((mw - 4.67) ** 2 + 0.341)) + F_a
+        b = -3.090 + 1.825 * mw + (1.206 / ((mw - 4.62) ** 2 + 0.263)) + F_b
+
+    elif 8.0 <= mw <= 10.0:
+        # Far-UV
+        c1 = [-1.073, -0.628, 0.137, -0.070]
+        c2 = [13.670, 4.257, -0.420, 0.374]
+        y = mw - 8.0
+        c1.reverse(), c2.reverse()
+        a, b = np.polyval(c1, y), np.polyval(c2, y)
+    else:
+        raise ValueError(
+            "The effective wavelength is {} [1/micron], beyond "
+            "the CCM model limit (10 [1/micron]).".format(mw)
+        )
+
+    return a, b
 
 
 def sample_imf(self, Nmets: int, Nages: int) -> list:
@@ -238,14 +300,13 @@ def sample_imf(self, Nmets: int, Nages: int) -> list:
       variety to the sampled IMF.
 
     """
-    print(f"Sampling selected IMF ({self.IMF_name})")
     inv_cdf = invTrnsfSmpl(self.IMF_name)
 
     st_dist_mass = []
     for i in range(Nmets):
         met_lst = []
         for j in range(Nages):
-            sampled_IMF = sampleInv(i+j + self.seed, self.max_mass, inv_cdf)
+            sampled_IMF = sampleInv(i + j + self.seed, self.max_mass, inv_cdf)
             met_lst.append(sampled_IMF)
         st_dist_mass.append(met_lst)
 
@@ -389,68 +450,6 @@ def mag_combine(m1: np.ndarray, m2: np.ndarray) -> np.ndarray:
     return mbin
 
 
-def ccm_model(mw: float) -> list:
-    """Cardelli, Clayton, and Mathis (1989 ApJ. 345, 245) model for extinction
-    coefficients with updated coefficients for near-UV from O'Donnell
-    (1994, ApJ, 422, 158).
-
-    ccm_coef = a + b / Rv
-
-    Implementation taken from:
-
-    http://idlastro.gsfc.nasa.gov/ftp/pro/astro/ccm_unred.pro
-
-    There appears to be an error in the Far-UV range in the original IDL
-    routine where the maximum inverse wavelength is 11 and it should be 10
-    according to Cardelli et al. 1989 (pag 251, Eq (5,a,b)).
-
-    """
-
-    if 0.3 <= mw < 1.1:
-        # Infrared.
-        a, b = 0.574 * (mw**1.61), -0.527 * (mw**1.61)
-
-    elif 1.1 <= mw < 3.3:
-        # Optical/NIR.
-        # Original coefficients from CCM89
-        # c1 = [1., 0.17699, -0.50447, -0.02427, 0.72085, 0.01979, -0.77530,
-        #       0.32999]
-        # c2 = [0., 1.41338, 2.28305, 1.07233, -5.38434, -0.62251, 5.30260,
-        #       -2.09002]
-        # New coefficients from O'Donnell (1994)
-        c1 = [1.0, 0.104, -0.609, 0.701, 1.137, -1.718, -0.827, 1.647, -0.505]
-        c2 = [0.0, 1.952, 2.908, -3.989, -7.985, 11.102, 5.491, -10.805, 3.347]
-        y = mw - 1.82
-        # Reverse because polyval starts from the highest degree.
-        c1.reverse(), c2.reverse()
-        a, b = np.polyval(c1, y), np.polyval(c2, y)
-
-    elif 3.3 <= mw < 8.0:
-        # Mid-UV
-        F_a, F_b = 0.0, 0.0
-        if mw >= 5.9:
-            y = mw - 5.9
-            F_a = -0.04473 * y**2 - 0.009779 * y**3
-            F_b = 0.2130 * y**2 + 0.1207 * y**3
-        a = 1.752 - 0.316 * mw - (0.104 / ((mw - 4.67) ** 2 + 0.341)) + F_a
-        b = -3.090 + 1.825 * mw + (1.206 / ((mw - 4.62) ** 2 + 0.263)) + F_b
-
-    elif 8.0 <= mw <= 10.0:
-        # Far-UV
-        c1 = [-1.073, -0.628, 0.137, -0.070]
-        c2 = [13.670, 4.257, -0.420, 0.374]
-        y = mw - 8.0
-        c1.reverse(), c2.reverse()
-        a, b = np.polyval(c1, y), np.polyval(c2, y)
-    else:
-        raise ValueError(
-            "The effective wavelength is {} [1/micron], beyond "
-            "the CCM model limit (10 [1/micron]).".format(mw)
-        )
-
-    return a, b
-
-
 def properModel(
     met_age_dict: dict, fix_params: dict, fit_params: dict
 ) -> tuple[float, float, float, float, float, float, float, int, int, int, int]:
@@ -490,10 +489,10 @@ def properModel(
         mh = min(len(par) - 1, np.searchsorted(par, model_full[0]))
         ml = mh - 1
 
-    if len(met_age_dict["a"]) == 1:
+    if len(met_age_dict["loga"]) == 1:
         al = ah = 0
     else:
-        par = met_age_dict["a"]
+        par = met_age_dict["loga"]
         ah = min(len(par) - 1, np.searchsorted(par, model_full[1]))
         al = ah - 1
 
@@ -529,7 +528,7 @@ def zaWAverage(theor_tracks, met_age_dict, m_ini_idx, z_model, a_model, ml, mh, 
     # The four points in the (z, age) grid that define the box that contains
     # the model value (z_model, a_model)
     z1, z2 = met_age_dict["met"][ml], met_age_dict["met"][mh]
-    a1, a2 = met_age_dict["a"][al], met_age_dict["a"][ah]
+    a1, a2 = met_age_dict["loga"][al], met_age_dict["loga"][ah]
     pts = np.array([(z1, a1), (z1, a2), (z2, a1), (z2, a2)])
 
     # Order: (z1, a1), (z1, a2), (z2, a1), (z2, a2)
@@ -639,7 +638,17 @@ def move_isochrone(isochrone, m_ini_idx, dm):
 
 
 def extinction(
-    ext_coefs, rand_norm, rand_unif, DR_distribution, m_ini_idx, Av, dr, Rv, isochrone
+    ext_law,
+    ext_coefs,
+    rand_norm,
+    rand_unif,
+    DR_distribution,
+    m_ini_idx,
+    binar_flag,
+    Av,
+    dr,
+    Rv,
+    isochrone,
 ):
     """
     Modifies magnitude and color(s) according to given values for the
@@ -649,6 +658,9 @@ def extinction(
     The distance modulus was applied before this function.
 
     isochrone = [mag, c1, (c2), .., Mini, mag_b, c1b, .., Mini_b]
+
+    For the CCMO model:
+
     ext_coefs = [mag_ec, c1_ec, ...]
 
     where:
@@ -670,6 +682,7 @@ def extinction(
                = (a12 + b12/Rv) * R_V * E(B-V)
     (m1 - m2)_obs = (m1 - m2)_int + E(m1 - m2)
     (m1 - m2)_obs = (m1 - m2)_int + (a12 + b12/Rv) * R_V * E(B-V)
+
     """
 
     if dr > 0.0:
@@ -689,29 +702,120 @@ def extinction(
         # Clip at 0
         Av = np.clip(Av + Av_dr, a_min=0, a_max=np.inf)
 
-    def colmove(ci):
-        Ex = (
-            (ext_coefs[ci][0][0] + ext_coefs[ci][0][1] / Rv)
-            - (ext_coefs[ci][1][0] + ext_coefs[ci][1][1] / Rv)
-        ) * Av
-        return Ex
+    if ext_law == "CCMO":
+        # Magnitude
+        ec_mag = ext_coefs[0][0] + ext_coefs[0][1] / Rv
+        # First color
+        ec_col1 = (ext_coefs[1][0][0] + ext_coefs[1][0][1] / Rv) - (
+            ext_coefs[1][1][0] + ext_coefs[1][1][1] / Rv
+        )
+    elif ext_law == "GAIADR3":
+        # If this model is used the first color is always expected to be BP-RP
+        # BP_RP = isochrone[1]
+        ec_mag, ec_col1 = dustapprox(isochrone[1], Av)
 
-    # Move magnitude
-    Ax_d = (ext_coefs[0][0] + ext_coefs[0][1] / Rv) * Av
-    isochrone[0] += Ax_d
-    # Move filters with binary data.
-    if isochrone.shape[0] > m_ini_idx + 1:
-        isochrone[m_ini_idx + 1] += Ax_d
+    Ax = ec_mag * Av
+    isochrone[0] += Ax
+    Ex1 = ec_col1 * Av
+    isochrone[1] += Ex1
 
-    # Move colors.
-    for ci in range(1, m_ini_idx):
-        Ex = colmove(ci)
-        isochrone[ci] += Ex
-        # Move colors with binary data.
-        if isochrone.shape[0] > m_ini_idx + 1:
-            isochrone[m_ini_idx + 1 + ci] += Ex
+    # Move binary data.
+    if binar_flag:
+        isochrone[m_ini_idx + 1] += Ax  # Magnitude
+        isochrone[m_ini_idx + 2] += Ex1  # First color
+
+    # Second color
+    if len(ext_coefs) > 2:
+        ec_col2 = (ext_coefs[2][0][0] + ext_coefs[2][0][1] / Rv) - (
+            ext_coefs[2][1][0] + ext_coefs[2][1][1] / Rv
+        )
+        Ex2 = ec_col2 * Av
+        isochrone[2] += Ex2
+        # Move color with binary data.
+        if binar_flag:
+            isochrone[m_ini_idx + 3] += Ex2
 
     return isochrone
+
+
+def dustapprox(X_, Av):
+    """
+    The 'coeffs' values are the main sequence values taken from:
+
+    https://www.cosmos.esa.int/web/gaia/edr3-extinction-law
+
+    The order of the coefficients is:
+
+    Intercept   X   X2  X3  A   A2  A3  XA  AX2 XA2
+
+    X_ == BP-RP
+    """
+    coeffs = {
+        "G": (
+            0.995969721536602,
+            -0.159726460302015,
+            0.0122380738156057,
+            0.00090726555099859,
+            -0.0377160263914123,
+            0.00151347495244888,
+            -2.52364537395142e-05,
+            0.0114522658102451,
+            -0.000936914989014318,
+            -0.000260296774134201,
+        ),
+        "BP": (
+            1.15363197483424,
+            -0.0814012991657388,
+            -0.036013023976704,
+            0.0192143585568966,
+            -0.022397548243016,
+            0.000840562680547171,
+            -1.31018008013549e-05,
+            0.00660124080271006,
+            -0.000882247501989453,
+            -0.000111215755291684,
+        ),
+        "RP": (
+            0.66320787941067,
+            -0.0179847164933981,
+            0.000493769449961458,
+            -0.00267994405695751,
+            -0.00651422146709376,
+            3.30179903473159e-05,
+            1.57894227641527e-06,
+            -7.9800898337247e-05,
+            0.000255679812110045,
+            1.10476584967393e-05,
+        ),
+    }
+
+    X_2 = X_**2
+    X_3 = X_**3
+    Av_2 = Av**2
+    Av_3 = Av**3
+
+    def ext_coeff(k):
+        """
+        https://www.cosmos.esa.int/web/gaia/edr3-extinction-law
+        """
+        # X   X2  X3  A   A2  A3  XA  AX2 XA2
+        ay = coeffs[k][0]
+        for i, Xk in enumerate([X_, X_2, X_3]):
+            ay += coeffs[k][1 + i] * Xk
+        for i, Ak in enumerate([Av, Av_2, Av_3]):
+            ay += coeffs[k][4 + i] * Ak
+
+        ay += (
+            coeffs[k][7] * X_ * Av
+            + coeffs[k][9] * X_ * Av_2  # This index not a mistake
+            + coeffs[k][8] * X_2 * Av
+        )
+        return ay
+
+    ec_G = ext_coeff("G")
+    ec_BPRP = ext_coeff("BP") - ext_coeff("RP")
+
+    return ec_G, ec_BPRP
 
 
 def cut_max_mag(isoch_moved, max_mag_syn):
@@ -902,11 +1006,13 @@ def generate(self, fit_params, plotflag=False):
 
     # Apply extinction correction
     isoch_extin = extinction(
+        self.ext_law,
         self.ext_coefs,
         self.rand_floats["norm"][0],
         self.rand_floats["unif"][0],
         self.DR_distribution,
         self.m_ini_idx,
+        self.binar_flag,
         av,
         dr,
         rv,